Add extendable loader runner mechanism
This introduces a setup that allows steps to be defined in a self contained
way. Simply create a file in the directory and inherit the corresponding
meta_type. The loader finds the potential steps and the runner then executes
them. I've also added some simple steps for running simple pipelines